Refactor permissions to be kept outside of redux and enable auto refresh
