+++ src/components/MAPI/__tests__/utils.ts
@@ -1,4 +1,8 @@
-import { determineRandomAZs, generateUID } from 'MAPI/utils';
+import {
+  determineRandomAZs,
+  generateUID,
+  getNamespaceFromOrgName,
+} from 'MAPI/utils';
 
 describe('mapi::utils', () => {
   describe('generateUID', () => {
@@ -85,6 +89,33 @@ describe('mapi::utils', () => {
     it('returns the maximum number of availability zones possible, even if the available ones are less than we need', () => {
       const azs = determineRandomAZs(3, ['1', '2']);
       expect(azs).toStrictEqual(['1', '2']);
+    });
+  });
+
+  describe('getNamespaceFromOrgName', () => {
+    test.each`
+      orgName                                                                            | expected
+      ${''}                                                                              | ${''}
+      ${'%_#!'}                                                                          | ${''}
+      ${'someorg'}                                                                       | ${'org-someorg'}
+      ${'someOrg'}                                                                       | ${'org-someorg'}
+      ${'some-org'}                                                                      | ${'org-some-org'}
+      ${'some-Org'}                                                                      | ${'org-some-org'}
+      ${'some_org'}                                                                      | ${'org-some-org'}
+      ${'some_Org'}                                                                      | ${'org-some-org'}
+      ${'some____Org'}                                                                   | ${'org-some-org'}
+      ${'some$Org'}                                                                      | ${'org-some-org'}
+      ${'some-random%org'}                                                               | ${'org-some-random-org'}
+      ${'some-org-'}                                                                     | ${'org-some-org'}
+      ${'-some-org-'}                                                                    | ${'org-some-org'}
+      ${'-some-org----'}                                                                 | ${'org-some-org'}
+      ${'-some-org%'}                                                                    | ${'org-some-org'}
+      ${'-some-random123_org'}                                                           | ${'org-some-random123-org'}
+      ${'some-random_org401'}                                                            | ${'org-some-random-org401'}
+      ${'some-random-org-with-random-keys-in-a-random-order-just-to-get-a-lot-of-chars'} | ${'org-some-random-org-with-random-keys-in-a-random-order-just-to'}
+    `(`computes namespace from '$orgName'`, ({ orgName, expected }) => {
+      const result = getNamespaceFromOrgName(orgName);
+      expect(result).toEqual(expected);
     });
   });
 });
