+++ src/components/MAPI/permissions/types.ts
@@ -1,0 +1,15 @@
+export type PermissionVerb =
+  | 'get'
+  | 'watch'
+  | 'list'
+  | 'create'
+  | 'update'
+  | 'patch'
+  | 'delete'
+  | '*'
+  | string;
+
+export interface INamespacePermissions
+  extends Record<string, PermissionVerb[]> {}
+
+export interface IPermissionMap extends Record<string, INamespacePermissions> {}
