+++ src/stores/main/constants.ts
@@ -20,5 +20,3 @@ export const REQUEST_PASSWORD_RECOVERY_TOKEN_REQUEST =
   'REQUEST_PASSWORD_RECOVERY_TOKEN_REQUEST';
 export const VERIFY_PASSWORD_RECOVERY_TOKEN = 'VERIFY_PASSWORD_RECOVERY_TOKEN';
 export const SET_NEW_PASSWORD = 'SET_NEW_PASSWORD';
-
-export const SET_PERMISSIONS = 'SET_PERMISSIONS';
