+++ src/stores/main/reducer.ts
@@ -9,7 +9,6 @@ import {
   LOGOUT_SUCCESS,
   REFRESH_USER_INFO_ERROR,
   REFRESH_USER_INFO_SUCCESS,
-  SET_PERMISSIONS,
 } from 'stores/main/constants';
 import { IMainState, MainActions } from 'stores/main/types';
 import {
@@ -28,7 +27,6 @@ const initialState = (): IMainState => ({
   selectedClusterID: null,
   firstLoadComplete: false,
   loggedInUser: fetchUserFromStorage(),
-  permissions: {},
 });
 
 const makeMainReducer = () => {
@@ -76,11 +74,6 @@ const makeMainReducer = () => {
 
         case CLUSTER_SELECT:
           draft.selectedClusterID = action.clusterID;
-
-          break;
-
-        case SET_PERMISSIONS:
-          draft.permissions = action.permissions;
 
           break;
       }
