+++ src/stores/main/types.ts
@@ -14,7 +14,6 @@ import {
   REFRESH_USER_INFO_SUCCESS,
   REQUEST_PASSWORD_RECOVERY_TOKEN_REQUEST,
   SET_NEW_PASSWORD,
-  SET_PERMISSIONS,
   VERIFY_PASSWORD_RECOVERY_TOKEN,
 } from 'stores/main/constants';
 
@@ -25,7 +24,6 @@ export enum LoggedInUserTypes {
 
 export interface IMainState {
   loggedInUser: ILoggedInUser | null;
-  permissions: IPermissionMap;
   selectedOrganization: string | null;
   firstLoadComplete: boolean;
   selectedClusterID: string | null;
@@ -102,11 +100,6 @@ export interface IMainSetNewPasswordAction {
   type: typeof SET_NEW_PASSWORD;
 }
 
-export interface IMainSetPermissionsAction {
-  type: typeof SET_PERMISSIONS;
-  permissions: IPermissionMap;
-}
-
 export type MainActions =
   | IMainSelectClusterAction
   | IGlobalLoadRequestAction
@@ -123,5 +116,4 @@ export type MainActions =
   | IMainRefreshUserInfoSuccessAction
   | IMainRequestPasswordRecoveryTokenAction
   | IMainVerifyPasswordRecoveryTokenAction
-  | IMainSetNewPasswordAction
-  | IMainSetPermissionsAction;
+  | IMainSetNewPasswordAction;
